[ADD] Component for ticket page #256
